requirements.txt
matplotlib
numpy
networkx
gitpython
pygraphviz